move party chat event into player chat manager
