+++ src/main/java/at/hannibal2/skyhanni/data/hypixel/chat/PartyChatEvent.kt
@@ -1,0 +1,10 @@
+package at.hannibal2.skyhanni.data.hypixel.chat
+
+import at.hannibal2.skyhanni.events.LorenzChatEvent
+import at.hannibal2.skyhanni.events.LorenzEvent
+
+class PartyChatEvent(
+    val author: String,
+    val text: String,
+    val trigger: LorenzChatEvent,
+) : LorenzEvent()
