--- src/main/java/at/hannibal2/skyhanni/events/PartyChatEvent.kt
@@ -1,7 +1,0 @@
-package at.hannibal2.skyhanni.events
-
-data class PartyChatEvent(
-    val author: String,
-    val text: String,
-    val trigger: LorenzChatEvent,
-) : LorenzEvent()
